+++ src/Collections/SavedTokens.php
@@ -64,7 +64,7 @@ class SavedTokens extends TerminusCollection implements ConfigAwareInterface, Da
      */
     public function deleteAll()
     {
-        foreach ($this->getMembers() as $token) {
+        foreach ($this->all() as $token) {
             $token->delete();
         }
     }
@@ -74,7 +74,7 @@ class SavedTokens extends TerminusCollection implements ConfigAwareInterface, Da
      */
     public function getData()
     {
-        if (empty($this->data)) {
+        if (empty(parent::getData())) {
             $tokens = [];
             foreach ($this->getDataStore()->keys() as $key) {
                 $tokens[] = (object)$this->getDataStore()->get($key);